migrate mixins and compat
