+++ src/main/java/su/terrafirmagreg/tfcambiental/api/EntityTemperatureProvider.java
@@ -8,6 +8,7 @@ import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.phys.AABB;
 
+import su.terrafirmagreg.core.common.entity.slime.TFGSlime;
 import su.terrafirmagreg.tfcambiental.TFCAmbiental;
 import su.terrafirmagreg.tfcambiental.modifier.TempModifier;
 import su.terrafirmagreg.tfcambiental.modifier.TempModifierStorage;
@@ -48,5 +49,26 @@ public interface EntityTemperatureProvider {
             return TempModifier.defined("cold_entity", -1 * change.get(), 0);
         }
         return TempModifier.none();
+    }
+
+    static Optional<TempModifier> getEntityTempModifier(Player player) {
+        float change = 0F;
+
+        for (Entity entity : player.level().getEntitiesOfClass(Entity.class,
+                new AABB(player.blockPosition()).inflate(5.0D, 2.0D, 5.0D))) {
+            if (entity.getType().is(TFCAmbiental.HOT_ENTITIES)) {
+                change += 1F;
+            } else if (entity.getType().is(TFCAmbiental.COLD_ENTITIES)) {
+                change -= 1F;
+            } else if (entity instanceof TFGSlime slime) {
+                change += slime.getAmbientalTemperature();
+            }
+        }
+
+        if (change == 0F) {
+            return Optional.empty();
+        }
+
+        return Optional.of(new TempModifier("tfg_entity_temperature", change, 0F));
     }
 }
