+++ src/main/java/su/terrafirmagreg/tfcambiental/capability/TemperatureCapability.java
@@ -10,9 +10,12 @@ import net.minecraft.core.registries.Registries;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.util.Mth;
 import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.common.capabilities.CapabilityToken;
 import net.minecraftforge.common.capabilities.ICapabilitySerializable;
@@ -102,26 +105,65 @@ public class TemperatureCapability implements ICapabilitySerializable<CompoundTa
 
     public void evaluateModifiers() {
         this.clearModifiers();
-        ItemTemperatureProvider.evaluateAll(this.player, this.modifiers);
-        EnvironmentalTemperatureProvider.evaluateAll(this.player, this.modifiers);
-        BlockTemperatureProvider.evaluateAll(this.player, this.modifiers);
-        BlockEntityTemperatureProvider.evaluateAll(this.player, this.modifiers);
-        EquipmentTemperatureProvider.evaluateAll(this.player, this.modifiers);
-        EntityTemperatureProvider.evaluateAll(this.player, this.modifiers);
+
+        boolean fullyInsulated = EquipmentTemperatureProvider.isFullyInsulated(this.player);
+
+        if (this.player.level().dimension() == Level.NETHER) {
+            // unroll EnvironmentalTemperatureProvider.evaluateAll but skip EquipmentTemperatureProvider::handleSunlightCap
+            CuriosApi.getCuriosHelper().getEquippedCurios(player).ifPresent(c -> {
+                for (int i = 0; i < c.getSlots(); i++) {
+                    ItemStack stack = c.getStackInSlot(i);
+                    for (var fn : AmbientalRegistry.EQUIPMENT) {
+                        this.modifiers.add(fn.getModifier(player, stack));
+                    }
+                }
+            });
+            this.modifiers.add(EquipmentTemperatureProvider.handleClothes(player, player.getItemBySlot(EquipmentSlot.HEAD)));
+            this.modifiers.add(EquipmentTemperatureProvider.handleClothes(player, player.getItemBySlot(EquipmentSlot.CHEST)));
+            this.modifiers.add(EquipmentTemperatureProvider.handleClothes(player, player.getItemBySlot(EquipmentSlot.LEGS)));
+            this.modifiers.add(EquipmentTemperatureProvider.handleClothes(player, player.getItemBySlot(EquipmentSlot.FEET)));
+
+            // unroll EnvironmentalTemperatureProvider.evaluateAll but skip select checks
+            this.modifiers.add(EnvironmentalTemperatureProvider.handleGeneralTemperature(player));
+            // this.modifiers.add(EnvironmentalTemperatureProvider.handleTimeOfDay(player));
+            // this.modifiers.add(EnvironmentalTemperatureProvider.handleShade(player));
+            // this.modifiers.add(EnvironmentalTemperatureProvider.handleCozy(player));
+            this.modifiers.add(EnvironmentalTemperatureProvider.handleThirst(player));
+            this.modifiers.add(EnvironmentalTemperatureProvider.handleFood(player));
+            this.modifiers.add(EnvironmentalTemperatureProvider.handleDiet(player));
+            this.modifiers.add(EnvironmentalTemperatureProvider.handleFire(player));
+            this.modifiers.add(EnvironmentalTemperatureProvider.handleWater(player));
+            // this.modifiers.add(EnvironmentalTemperatureProvider.handleRain(player));
+            // this.modifiers.add(EnvironmentalTemperatureProvider.handleWind(player));
+            this.modifiers.add(EnvironmentalTemperatureProvider.handleSprinting(player));
+            // this.modifiers.add(EnvironmentalTemperatureProvider.handleUnderground(player));
+            this.modifiers.add(EnvironmentalTemperatureProvider.handleWetness(player));
+        } else {
+            EquipmentTemperatureProvider.evaluateAll(this.player, this.modifiers);
+            EnvironmentalTemperatureProvider.evaluateAll(this.player, this.modifiers);
+        }
+        if (!fullyInsulated) {
+            ItemTemperatureProvider.evaluateAll(this.player, this.modifiers);
+            BlockTemperatureProvider.evaluateAll(this.player, this.modifiers);
+            BlockEntityTemperatureProvider.evaluateAll(this.player, this.modifiers);
+            this.modifiers.add(EntityTemperatureProvider.getEntityTempModifier(this.player));
+        }
         this.modifiers.keepOnlyNEach(3);
 
-        this.target = this.modifiers.getTargetTemperature();
         this.potency = this.modifiers.getTotalPotency();
+        this.target = this.modifiers.getTargetTemperature();
         this.targetWetness = this.modifiers.getTargetWetness();
 
-        if (this.target > this.temperature && this.temperature > TFCAmbientalConfig.COMMON.hotThreshold.get().floatValue()) {
-            this.potency /= this.potency;
-        }
-        if (this.target < this.temperature && this.temperature < TFCAmbientalConfig.COMMON.coolThreshold.get().floatValue()) {
-            this.potency /= this.potency;
+        if ((this.target > this.temperature && this.temperature > TFCAmbientalConfig.COMMON.hotThreshold.get().floatValue())
+                || (this.target < this.temperature && this.temperature < TFCAmbientalConfig.COMMON.coolThreshold.get().floatValue())) {
+            this.potency = 1f;
         }
 
         this.potency = Math.max(1f, this.potency);
+
+        if (fullyInsulated) {
+            this.target = Mth.clamp(this.target, 5f, 25f);
+        }
     }
 
     public float getTargetTemperature() {
